+++ Test3/Assembly-CSharp.csproj
@@ -46,10 +46,10 @@
 	</Reference>
   </ItemGroup>
   <ItemGroup>
-     <Compile Include="Assets\Scripts\Enemy\BaseballWolf\BaseballWolfMovement.cs" />
+     <Compile Include="Assets\Scripts\Enemy\BaseballWolf\BaseballWolf.cs" />
      <Compile Include="Assets\Scripts\Enemy\EnemySpawner.cs" />
      <Compile Include="Assets\Scripts\Enemy\FootballWolf\FootballScript.cs" />
-     <Compile Include="Assets\Scripts\Enemy\FootballWolf\FootballWolfMovement.cs" />
+     <Compile Include="Assets\Scripts\Enemy\FootballWolf\FootballWolf.cs" />
      <Compile Include="Assets\Scripts\Enemy\General\Avoid.cs" />
      <Compile Include="Assets\Scripts\Enemy\General\Enemy.cs" />
      <Compile Include="Assets\Scripts\Enemy\Mikey\Movement\MikeyMovement.cs" />
