+++ Test3/Assets/Scripts/Enemy/FootballWolf/FootballWolf.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 86538856091a9de48a22ed7d86f98f72
-timeCreated: 1444023832
+guid: 8324ba90b3f438b4380b908683451e37
+timeCreated: 1447445176
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
